make code compilable with local GDAL lib
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@3012 d0c010d8-bb82-477e-a075-112140230451

+++ src/CMapQMAPExport.cpp
@@ -556,7 +556,7 @@ void CMapQMAPExport::slotFinished1( int exitCode, QProcess::ExitStatus status)
                 textBrowser->setTextColor(Qt::black);
                 textBrowser->append(job.tarFilename+"\n");
                 job.srcFilename=job.tarFilename;
-                quint64 div = (quint64)sqrt((job.width*job.height)/MAX_MYNAV);
+                quint64 div = (quint64)sqrt((double)(job.width*job.height)/MAX_MYNAV);
                 div++;
                 int tilesX = job.width/div;
                 int tilesY = job.height/div;
